updated ms repo

+++ milestones/summer24--sft221-nbb-2/SourceCode/Milestone/Milestone_main/data.h
@@ -2,9 +2,6 @@
 #define DATA_H
 #include "mapping.h"
 
-
-
-
 /**
  * Struct for a Package or Shipment
  * @param Point destination - points on the map corresponding to package destination
@@ -28,7 +25,7 @@ struct Package {
 struct Truck {
 	struct Point* currentPosition;
     double currentWeight;
-    int currentVolume;
+    double currentVolume;
     char routeSymbol;
     struct Route* route;
     int routeLength;
